backend/apps/assistant_settings/serializers.py
from rest_framework import serializers
from .models import BaseModel

class BaseModelSerializer(serializers.ModelSerializer):
    class Meta:
        model = BaseModel
        fields = '__all__'
        
