backend/requirements.txt
streamlit 
tqdm
boto3 
django 
django-simpleui 
djangorestframework 
djangorestframework-simplejwt 
drf-yasg
pymysql
pycryptodomex
aiohttp
uvicorn